requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
tensorflow
python-multipart
Pillow
streamlit